fix: add waiting loops for contract events in sdk tests
