+++ sdk/tests/guarantee.rs
@@ -1,14 +1,13 @@
-use anyhow::bail;
 use rust_sdk_4mica::{
-    Address, Client, ConfigBuilder, PaymentGuaranteeRequestClaims, SigningScheme, U256,
-    client::recipient::RecipientClient, error::VerifyGuaranteeError,
+    Client, ConfigBuilder, PaymentGuaranteeRequestClaims, SigningScheme, U256,
+    error::VerifyGuaranteeError,
 };
 use serial_test::serial;
-use std::time::{Duration, Instant};
+use std::time::Duration;
 
 mod common;
 
-use crate::common::ETH_ASSET_ADDRESS;
+use crate::common::{ETH_ASSET_ADDRESS, wait_for_collateral_increase};
 
 #[tokio::test]
 #[serial]
@@ -299,40 +298,4 @@ async fn test_multiple_guarantees_increment_req_id() -> anyhow::Result<()> {
     assert_eq!(guarantees[1].req_id, parsed_second.req_id);
 
     Ok(())
-}
-
-async fn wait_for_collateral_increase(
-    recipient_client: &RecipientClient,
-    user_address: &str,
-    asset_address: Address,
-    starting_total: U256,
-    increase_by: U256,
-) -> anyhow::Result<()> {
-    let poll_interval = Duration::from_millis(500);
-    let timeout = Duration::from_secs(60);
-    let start = Instant::now();
-    let user_address = user_address.to_string();
-    let asset_address = asset_address.to_string();
-    let target_total = starting_total + increase_by;
-    let mut last_total = starting_total;
-
-    loop {
-        if let Some(balance) = recipient_client
-            .get_user_asset_balance(user_address.clone(), asset_address.clone())
-            .await?
-        {
-            last_total = balance.total;
-            if last_total >= target_total {
-                return Ok(());
-            }
-        }
-
-        if start.elapsed() > timeout {
-            bail!(
-                "timed out waiting for collateral increase to {target_total:?} for user {user_address}, last observed total {last_total:?}"
-            );
-        }
-
-        tokio::time::sleep(poll_interval).await;
-    }
 }
