+++ sdk/tests/remunerate.rs
@@ -6,7 +6,7 @@ use rust_sdk_4mica::{
 };
 use std::{str::FromStr, time::Duration};
 
-use crate::common::ETH_ASSET_ADDRESS;
+use crate::common::{ETH_ASSET_ADDRESS, wait_for_collateral_increase};
 
 mod common;
 
@@ -101,6 +101,12 @@ async fn test_recipient_remuneration() -> anyhow::Result<()> {
     let eth_asset_before =
         common::extract_asset_info(&user_info, ETH_ASSET_ADDRESS).expect("ETH asset not found");
 
+    let core_total_before = recipient_client
+        .recipient
+        .get_user_asset_balance(user_address.clone(), ETH_ASSET_ADDRESS.to_string())
+        .await?
+        .map(|info| info.total)
+        .unwrap_or(U256::ZERO);
     let deposit_amount = U256::from(2_000_000_000_000_000_000u128); // 2 ETH
     let _receipt = user_client.user.deposit(deposit_amount, None).await?;
 
@@ -112,7 +118,14 @@ async fn test_recipient_remuneration() -> anyhow::Result<()> {
         eth_asset_before.collateral + deposit_amount
     );
 
-    tokio::time::sleep(Duration::from_secs(2)).await;
+    wait_for_collateral_increase(
+        &recipient_client.recipient,
+        &user_address,
+        ETH_ASSET_ADDRESS,
+        core_total_before,
+        deposit_amount,
+    )
+    .await?;
 
     let tab_id = recipient_client
         .recipient
@@ -225,10 +238,23 @@ async fn test_double_remuneration_fails() -> anyhow::Result<()> {
     let recipient_client = Client::new(recipient_config).await?;
     log_signature_environment("double:recipient", &recipient_config_clone).await?;
 
+    let core_total_before = recipient_client
+        .recipient
+        .get_user_asset_balance(user_address.clone(), ETH_ASSET_ADDRESS.to_string())
+        .await?
+        .map(|info| info.total)
+        .unwrap_or(U256::ZERO);
     let deposit_amount = U256::from(2_000_000_000_000_000_000u128); // 2 ETH
     let _receipt = user_client.user.deposit(deposit_amount, None).await?;
 
-    tokio::time::sleep(Duration::from_secs(2)).await;
+    wait_for_collateral_increase(
+        &recipient_client.recipient,
+        &user_address,
+        ETH_ASSET_ADDRESS,
+        core_total_before,
+        deposit_amount,
+    )
+    .await?;
 
     let tab_id = recipient_client
         .recipient
